requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
python-dotenv
flask-cors
Flask-JWT-Extended
redis
